HW_4 added RestTemplate (some responses are String type and some are Json, because of exception while converting content type. And bad RestTemplateErrorHandler because I can't change appropriate status code)

+++ src/main/java/com/epam/springshop/service/impl/ProductServiceImpl.java
@@ -56,6 +56,10 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public void deleteProduct(Long obj) {
         log.info(String.format("%s : method ==> deleteProduct(%s)", this.getClass().getName(),obj));
+        Product product = productRepo.read(obj);
+        if(product == null){
+            throw new ProductNotFoundException();
+        }
         productRepo.delete(obj);
     }
 }
